set contextClassLoader to testClassLoader before loading config.
Together with
https://github.com/jdbeutel/spock/commit/71d04c503b06f04b9ad0caa9f7f3d9d5a1a5f0b5
this seems to fix http://code.google.com/p/spock/issues/detail?id=184
for me.  However, I do not know if it will cause other problems,
nor even how to get this project to run the new test using the
required changes in spock-core (to confirm it passes).